Normalize interface field types

+++ src/gd.h
@@ -5,16 +5,16 @@
 typedef long (*read_func_t)(int fd, char *buf, long count);
 
 typedef struct {
-    int status;        // current decoder status
-    int width;         // image height
-    int height;        // image height
+    uint8_t status;        // current decoder status
+    uint16_t width;         // image height
+    uint16_t height;        // image height
     bool gct;          // global color table present
     uint8_t gctb;      // global color table bits
 } gd_info_t;
 
 typedef struct {
-    int request;
-    int status;
+    uint8_t request;
+    uint8_t status;
     uint16_t width;
     uint16_t height;
     uint32_t *pixels;  // row-major array of pixel color table indices
